[WIP] TEMPORARILY drop chromium_src override for notification_platform_bridge_mac.mm
notification_platform_bridge_mac.mm has been migrated to C++ now and rewritten
entirely without relying on NSObjects (e.g. NSUserNotification), meaning that we
need to re-assess whether we need to keep having local modifications along the
lines of those introduced in commit 89c25f1c [1] in the context of PR #7516 [2].

If we want to keep such changes around, we'd need to reimplement the changes in
C++, so for now we're just dropping the notification_platform_bridge_mac.mm to
keep the build going, and dropping the associated patch.

[1] https://github.com/brave/brave-core/commit/89c25f1c
[2] https://github.com/brave/brave-core/pull/7516

Chromium change:

https://chromium.googlesource.com/chromium/src/+/83c4fd04ac6f669d41c97b2c4088e63daaad87af

commit 83c4fd04ac6f669d41c97b2c4088e63daaad87af
Author: Richard Knoll <knollr@chromium.org>
Date:   Fri Jul 23 10:12:19 2021 +0000

    Rename notification files from .mm to .cc

    These files are no longer written in Obj-C so renaming them to standard
    C++ conventions. Also makes sure to match filenames with the class names
    after recent cleanup CLs. Mechanical renaming CL only, no functional
    changes.

    Bug: 1170731